+++ java-client/src/main/java/io/sphere/client/facets/DateTimeRangeFacet.java
@@ -1,8 +1,12 @@
 package io.sphere.client.facets;
 
+import io.sphere.client.model.SearchResult;
 import io.sphere.client.model.facets.DateTimeRangeFacetItem;
+import io.sphere.client.model.facets.DateTimeRangeFacetResult;
 
-/** Marker interface for date time range facets. */
+/** Datetime range facet component. */
 public interface DateTimeRangeFacet extends Facet<DateTimeRangeFacetItem> {
+    /** Returns the results for this facet.  */
+    public <T> DateTimeRangeFacetResult getResult(SearchResult<T> searchResult);
 }