+++ java-client/src/main/java/io/sphere/client/model/facets/FacetResult.java
@@ -3,11 +3,11 @@ package io.sphere.client.model.facets;
 import org.codehaus.jackson.annotate.JsonSubTypes;
 import org.codehaus.jackson.annotate.JsonTypeInfo;
 
-/** Information about a facet, as part of {@link io.sphere.client.model.SearchResult}. */
+/** Calculated results for a facet, as part of {@link io.sphere.client.model.SearchResult}. */
 @JsonTypeInfo(use=JsonTypeInfo.Id.NAME, include=JsonTypeInfo.As.PROPERTY, property="type")
 @JsonSubTypes({
         @JsonSubTypes.Type(value = TermFacetResult.class, name = "terms"),
-        @JsonSubTypes.Type(value = RangeFacetResult.class, name = "range")
+        @JsonSubTypes.Type(value = RangeFacetResultRaw.class, name = "range")
 })
 public interface FacetResult {
 }